backend/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# OCR et traitement d'images
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
pdf2image==1.17.0

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentification
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Export de données
pandas==2.2.0
